+++ kcstog.php
@@ -11,15 +11,17 @@ if(!checkState(8)) { # keyed off the downlight at the far end of the kitchen. it
     saveCTState(8);
     saveBriState(34); # Hue filament bulb has no HS or CT properties, but we need Bri (just in case)
     saveOnState(27); # The INNR smart-plug is strictly on/off - no other properties exist
+    saveOnState(41); # Fairy lights
 
     oneOn(27);
+    oneOn(41);
     setHSState('true',4,38,7676,199);
     setCTState('true', 6,127,400);
     setCTState('true', 7,254,400);
     setCTState('true', 8,127,400);
     setLevel(34,38);
 } else { # we're in-progress, so just restore previous states for these
-    $lightlist = array(4,6,7,8,27,34);
+    $lightlist = array(4,6,7,8,27,34,41);
 
     foreach($lightlist as $id) {
         restoreState($id);